# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Inrave API",
        "description": "This is the API documentation for Inrave project.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://a1050418.xsph.ru/",
            "description": "Local API server"
        }
    ],
    "paths": {
...
        "/api/trainers": {
            "get": {
                "tags": [
                    "Trainers"
                ],
                "summary": "Get a list of trainers",
                "operationId": "116f689cd23c33d0d52057ab975e4a18",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Filter trainers by first or last name",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "service_id",
                        "in": "query",
                        "description": "Filter trainers by service ID",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Trainer"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Trainers"
                ],
                "summary": "Create a new trainer",
                "operationId": "02af24b16f7d770be5454aa8b0d33fbd",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "first_name",
                                    "last_name",
                                    "description"
                                ],
                                "properties": {
                                    "first_name": {
                                        "type": "string",
                                        "example": "John"
                                    },
                                    "last_name": {
                                        "type": "string",
                                        "example": "Doe"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Experienced boxing trainer."
                                    },
                                    "photo": {
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "services": {
                                        "type": "array",
                                        "items": {
                                            "type": "integer",
                                            "example": 1
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Trainer created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Trainer"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
        "/api/trainers/{id}": {
            "get": {
                "tags": [
                    "Trainers"
                ],
                "summary": "Get a specific trainer by ID",
                "operationId": "dff48a85414a68482bfd49897b6a8656",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Trainer"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Trainer not found"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Trainers"
                ],
                "summary": "Update an existing trainer",
                "operationId": "ab2c71f48ea640c221b3fa20473989a2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "first_name": {
                                        "type": "string",
                                        "example": "John"
                                    },
                                    "last_name": {
                                        "type": "string",
                                        "example": "Doe"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Updated description"
                                    },
                                    "photo": {
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "services": {
                                        "type": "array",
                                        "items": {
                                            "type": "integer",
                                            "example": 1
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Trainer updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Trainer"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Trainer not found"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Trainers"
                ],
                "summary": "Delete a trainer by ID",
                "operationId": "f34b04bc84320ad785fba6b1cbc37dbf",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Trainer deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Trainer deleted successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Trainer not found"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Service": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "title": {
                        "type": "string",
                        "example": "Boxing Training"
                    },
                    "description": {
                        "type": "string",
                        "example": "Intensive boxing training sessions."
                    },
                    "price": {
                        "type": "number",
                        "format": "float",
                        "example": "99.99"
                    },
                    "image": {
                        "type": "string",
                        "example": "services/image.jpg",
                        "nullable": true
                    },
                    "trainers": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Trainer"
                        }
                    }
                },
                "type": "object"
            },
            "Trainer": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "first_name": {
                        "type": "string",
                        "example": "John"
                    },
                    "last_name": {
                        "type": "string",
                        "example": "Doe"
                    },
                    "description": {
                        "type": "string",
                        "example": "Experienced boxing trainer."
                    },
                    "photo": {
                        "type": "string",
                        "example": "trainers/photo.jpg",
                        "nullable": true
                    },
                    "services": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Service"
                        }
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "BearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}